Fix: Refactor Project Structure And Generate Same Id Bug Fix (#29)
[Fix] Refactor Project Structure And Generate Same Id Bug Fix
