+++ src/Routes/Router.js
@@ -6,7 +6,7 @@ import HomePage from "../Pages/Home/HomePage";
 import PageNotFound from "../Pages/PageNotFound/PageNotFound";
 import FindPasswordContainer from "../Pages/FindPassword/FindPasswordContainer";
 import InterviewConatiner from "../Pages/Interview/InterviewConatiner";
-import PostContainer from "../Pages/Post/PostPage/PostContainer";
+import PostPage from "../Pages/Post/PostPage/PostPage";
 import MyPostsContainer from "../Pages/MyPost/MyPostsContainer";
 import MyScrapContainer from "../Pages/MyScrap/MyScrapContainer";
 import ScrapInterviewContainer from "../Pages/ScrapInterview/ScrapInterviewContainer";
@@ -21,7 +21,7 @@ const Router = () => {
       <Route path="/select" element={<RegisterSelect />} />
       <Route path="/signup" element={<RegisterEmailContainer />} />
       <Route path="/find" element={<FindPasswordContainer />} />
-      <Route path="/post" element={<PostContainer />} />
+      <Route path="/post" element={<PostPage />} />
       <Route path="/my-posts" element={<MyPostsContainer />} />
       <Route path="/scrap" element={<MyScrapContainer />} />
       <Route path="/interview/:interview_id" element={<InterviewConatiner />} />
